refactor(identity): relocate gateway-assertion contract types to ServiceDefaults
Move GatewayAssertionPayload and GatewayAssertionToken from StarterApp.Api into
the shared StarterApp.ServiceDefaults.GatewayIdentity namespace (now public) so
the assertion signing/verification contract can be shared across projects. Shape
is unchanged (amr stays a first-class claim; no header hash). Base64UrlEncode is
now private to satisfy CA1055 on the public surface.
